// frontend/src/api/types.ts
export interface components {
    schemas: {
// ...
        Passage: {
            first_block_id: string;
            first_block_offset: number;
            last_block_id: string;
            last_block_offset: number;
            /** @description Kept for display and as a self-check. */
            text: string;
        };
// ...
    };
// ...
}